rename app logos to banners and mv to user-apps
