+++ ui/src/app/components/apps-panel/app-logo.tsx
@@ -7,10 +7,10 @@ export const AppLogo = (props: {
   appType: UIAppType;
   style: CSSProperties;
 }) => {
-  const { logo } = appAssets.find((aa) => aa.appType === props.appType);
+  const { banner } = appAssets.find((aa) => aa.appType === props.appType);
   return (
     <img
-      src={logo}
+      src={banner}
       alt={props.appType}
       aria-label={props.appType}
       style={props.style}
