+++ ui/src/app/components/help-sidebar-icons.tsx
@@ -693,6 +693,9 @@ export const HelpSidebarIcons = (props: HelpSidebarIconsProps) => {
     if (config.enableRStudioGKEApp) {
       keys.push(rstudioConfigIconId);
     }
+    if (config.enableSasGKEApp) {
+      keys.push(sasConfigIconId);
+    }
     keys.push('runtimeConfig', 'terminal');
   }
