+++ ui/src/assets/user-apps/Jupyter-icon.png
